refactor: optimize icons generation and export
The script `generate-icons.mjs` has been refactored to optimize the generation and export of icons data. The `simple-icons` library is used to create an object containing the SVG and hex values for each icon. The generated data is then written to `icons.ts` in the `src/assets` directory. This commit improves the efficiency of the icons generation process.